Conflicts:
	arch/arm/Kconfig
	arch/arm64/kernel/traps.c
	arch/arm64/mm/fault.c
	drivers/hwtracing/coresight/coresight-etm4x.c
	drivers/hwtracing/coresight/coresight-tmc.c
	drivers/hwtracing/stm/Kconfig
	drivers/hwtracing/stm/core.c
	drivers/mmc/card/block.c
	drivers/mmc/core/mmc.c
	drivers/staging/android/sync.c
	drivers/usb/gadget/configfs.c
	drivers/usb/gadget/function/u_serial.c
	drivers/video/fbdev/goldfishfb.c
	kernel/events/ring_buffer.c
	kernel/sched/fair.c
	mm/vmstat.c
	net/wireless/nl80211.c
	sound/core/compress_offload.c

+++ drivers/hwtracing/stm/core.c
@@ -403,11 +403,13 @@ static int stm_char_release(struct inode *inode, struct file *file)
 
 	stm_output_free(stm, &stmf->output);
 
+	stm_output_free(stmf->stm, &stmf->output);
+
 	/*
 	 * matches the stm_char_open()'s
 	 * class_find_device() + try_module_get()
 	 */
-	stm_put_device(stm);
+	stm_put_device(stmf->stm);
 	kfree(stmf);
 
 	return 0;
@@ -892,18 +894,8 @@ unlock:
 	spin_unlock(&src->link_lock);
 	spin_unlock(&stm->link_lock);
 
-	/*
-	 * Call the unlink callbacks for both source and stm, when we know
-	 * that we have actually performed the unlinking.
-	 */
-	if (!ret) {
-		if (src->data->unlink)
-			src->data->unlink(src->data);
-
-		if (stm->data->unlink)
-			stm->data->unlink(stm->data, src->output.master,
-					  src->output.channel);
-	}
+	if (!ret && src->data->unlink)
+		src->data->unlink(src->data);
 
 	return ret;
 }
